delete both input and result from database, and draw specific simulation when press on watch in history page

+++ services/database-service/src/controllers/simulationDelete.controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { SimulationResult } from "../models/simulationResult.model";
+import { SimulationInput } from "../models/simulationInput.model";
 
 export const deleteSimulation = async (
   req: Request<{ id: string }>,
@@ -7,12 +8,20 @@ export const deleteSimulation = async (
 ) => {
   try {
     const { id } = req.params;
+    console.log('id:',id);
 
-    const deleted = await SimulationResult.findByIdAndDelete(id);
+    const result = await SimulationResult. findById(id);
+    if(!result)
+      return res.status(404).json({error: 'sim not found'});
 
-    if (!deleted) {
+    const deletedResult = await SimulationResult.findByIdAndDelete(id);
+    if(!deletedResult)
       return res.status(404).json({ error: "Simulation not found" });
-    }
+
+    const inputID = result.simulationInputId;    
+    const deletedInput = await SimulationInput.findByIdAndDelete(inputID);
+    if(!deletedInput)
+      return res.status(404).json({ error: "Simulation not found" });
 
     res.status(200).json({ message: "Simulation deleted successfully" });
   } catch (err) {
